requirements.txt
pyyaml==6.0.1
boto3==1.24.96
